350_intersection_of_two_arrays_ii/intersection_of_two_arrays_ii_test.go
package intersection_of_two_arrays_ii

import (
	"reflect"
	"sort"
	"testing"
)

func Test_intersect(t *testing.T) {
	type args struct {
		nums1 []int
		nums2 []int
	}
	tests := []struct {
		name string
		args args
		want []int
	}{
		{
			name: "test1",
			args: args{
				nums1: []int{1, 2, 2, 1},
				nums2: []int{2, 2},
			},
			want: []int{2, 2},
		},
		{
			name: "test2",
			args: args{
				nums1: []int{4, 9, 5},
				nums2: []int{9, 4, 9, 8, 4},
			},
			want: []int{9, 4},
		},
		{
			name: "test3",
			args: args{
				nums1: []int{1},
				nums2: []int{},
			},
			want: []int{},
		},
		{
			name: "test4",
			args: args{
				nums1: []int{4, 9, 5, 9},
				nums2: []int{9, 4, 9, 8, 4},
			},
			want: []int{4, 9, 9},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			got := intersect(tt.args.nums1, tt.args.nums2)
			sort.Ints(got)
			sort.Ints(tt.want)
			if !reflect.DeepEqual(got, tt.want) {
				t.Errorf("intersect() = %v, want %v", got, tt.want)
			}
		})
	}
}
